+++ backend/app/schemas/statement.py
@@ -3,12 +3,13 @@ from pydantic import BaseModel
 from app.schemas.customer import CustomerRead
 from app.schemas.payment import PaymentRead
 from app.schemas.sale import SaleRead
+from decimal import Decimal
 
 
 class CustomerStatement(BaseModel):
     customer = CustomerRead
-    total_sales: float
-    total_payments: float
-    balance: float
+    total_sales: Decimal
+    total_payments: Decimal
+    balance: Decimal
     sales = list[SaleRead]
     payments = list[PaymentRead]
